util: update refcount code to ignore type info
GoRef and GoUnref now use a map of uint64 -> (referenced object, ref
count). This makes GoRef and GoUnref more flexible to changes in the
types of the objects that it is referencing.

Previously, the caller of GoUnref would have to ensure that the type of
the unreffed instance matched the type that was provided in the
corresponding GoRef call. In most cases, this was OK, but it proved
problematic in the implementation of Go<->Java channels.

Change-Id: I87a90c657697f3f00c1acb9d0bbe2c049b426803