util: add Push/PopLocalFrame calls
Typically JNI is responsible for freeing unused local references, but
this can only be guaranteed when Java is the originator of the native
method call. In certain situations (e.g. RPC handlers), calls originate
from within Go. In this case a Push/PopLocalFrame pair is required to
inform JNI that it is time to clear local references.

This fixes some "local reference table capacity exceeded" errors that we
were observing in Android.

Change-Id: If100f9d4b9a57b02c24ef3c7782a21e8cfe22868