impl/google/rpc: fix env cleanup
This CL cleans up env before making a potentially blocking channel
write. Previously, if the channel write blocked, the executing Go thread
would not be available for scheduling due to the fact that GetEnv()
calls runtime.LockOSThread. This could potentially starve Go for
threads.

Also now we explicitly call freeFunc() in the loop so that env is freed
after each loop iteration that uses it, rather than at the end of the
enclosing function.

Change-Id: Ie9cf18838c3a939424a8342e4b95d3448f5dc00d