v.io/x/jni: make GetEnv() re-entrant
Fixes the following bug:

This is our GetEnv() function that uses runtime.{Lock,Unlock}OSThread()
to prevent two threads ending up with the same *C.JNIEnv.

func GetEnv() (*C.JNIEnv, free func()) {
        runtime.LockOSThread()
	var env *C.JNIEnv
	if C.GetEnv(jVM, &env, C.JNI_VERSION_1_6) != C.JNI_OK {
		C.AttachCurrentThreadAsDaemon(jVM, &env, nil)
	}
        return env, func() {
		runtime.UnlockOSThread()
	}
}

This works well except when the same goroutine calls GetEnv() twice.  We were
doing this by a complicated Go->Java->Go->Java code-path, but the bottom line
is that the following was happening on the same goroutine:

GetEnv()
GetEnv()
free()      <-----  this one will unlock the thread
free()

So the thread got unlocked too early and we ended up hitting the two-threads
with the same *C.JNIEnv case.

Ideally, LockOSThread() could return a bool indicating whether it's already
been locked for the current goroutine.  We just get around this problem by
refcounting *env and unlocking only when the refcount goes to zero.

Change-Id: I9453f26e3bb86b00567736e71c4912b227537b21