v.io/x/jni: introduce type-sanity to the JNI code
This change was motivated by a bug where we were wrapping a C.jobject
reference in unsafe.Pointer, which caused Go to panic during garbage
collection (as C.jobject, a pointer type, often holds a local
reference which is a small int).

Instead, we now create types for JNI Env, Object, and Class, and pass
those around instead of passing unsafe.Poitners.

MultiPart: 1/2
Change-Id: I8d9aecc7aa36817c21194eaff33e30d31577bc18