TBR v.io/x/jni: Delete local references when done with them
This *may* fix the syncslides crashes observed by afergan/kash.  We can't
reproduce the problem in a test, however, but that could be due to a
more generous (or infinite) ref limit on desktops.

Together with sjr@, who correctly observed that we're overflowing a
local reference table, not global.

Note that it's still unclear if we really hit a case with 154 glob
results, or it may be something else going on.  But this CL *is*
the right thing to do regardless.

Change-Id: I1250616818358760965495a03d3c972989a857ed