jni: Remove generated IsZero logic.
The purpose of the IsZero logic was as a speed-up; the normal vom
zero-field encoding is slow, since it actually encodes fields and
then backtracks to throw away zero-valued fields.  However the
generated IsZero was actually significantly *slower*, for the
native types case, and possibly also the non-native types case.

In the future we will re-enable the IsZero logic for simple
cases (no native types, no typeobject or union).  And we may
enable IsZero for the more complicated cases, if we can come up
with a simple strategy that is actually faster that the default
backtracking implementation.

MultiPart: 2/6

Change-Id: Ib92f601c5275587b7f35adfaf3885c991105bfa9