jni: Add VDL codegen for VDLRead(vdl.Decoder) method of each type
This CL adds codegen support to the vdl compiler for the VDLRead
method, which is used for fast decoding via the vom.XDecoder.
The VDLRead approach is simpler and much faster than the old
vdl.Target interface approach, since it doesn't require any
reflection, and doesn't need to allocate new Targets.  Eventually
vdl.Target and all of its associated codegen will be removed.

The only feature missing from this CL is support for the Any
type, which needs both vdl.Value and vom.RawBytes support for
VDLRead.  Those features will be added in a later CL.

MultiPart: 2/5

Change-Id: I2bb470b7b03c04741ad8ab8185f9e9a50442065e